python_build_agent/requirements.txt
# Python Build Agent Dependencies

# LLM API Client
anthropic>=0.40.0
openai>=1.50.0

# Agent Framework (optional - we'll build custom)
# langgraph>=0.2.0
# langchain>=0.3.0

# Utility Libraries
python-dotenv>=1.0.0
rich>=13.7.0  # Beautiful terminal output
colorama>=0.4.6  # Cross-platform colored terminal
pydantic>=2.9.0  # Data validation

# Build tool detection
cmake>=3.29.0  # CMake Python wrapper (optional)
